[Projects] Simplify shared props & split Microsoft.Azure.EventHubs.Processor dependency
Overall changes:
- Utilized Directory.Build.props to DRY up projects (and misc. associated changes like one icon.png file, etc.)
- Creates a new DurableTask.Netherite.EventHubs project which is the only thing depending on Microsoft.Azure.EventHubs.Processor.
- Moves all EventHubs code possible to DurableTask.Netherite.EventHubs

The problem here is there are still a few refs in core classes we have to figure out:
- `ScalingMonitor` is calling `EventHubsConnections.GetQueuePositionsAsync` to see if work is left
- `FasterStorageProvider` is calling `EventHubsUtil.DeleteEventHubIfExistsAsync` to cleanup partition hubs to ensure queues start at zero

Other than those 2, the split is pretty clean here. I do recommend we fully drop `netcoreapp2.2` though - I assume based on comments and other cleanup that has happened it's explicitly here for a reason I don't understand, happy to understand that and maybe advise on some alternatives to clean up the build if we still need to keep it around.
