+++ DurableTask.Netherite.sln
@@ -24,6 +24,9 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PerformanceTests", "test\PerformanceTests\PerformanceTests.csproj", "{DD1E1B3F-4FA2-4F3A-9AE1-6B2A0B864AAF}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{D33AB157-04B9-4BAD-B580-C3C87C17828C}"
+	ProjectSection(SolutionItems) = preProject
+		src\Directory.Build.props = src\Directory.Build.props
+	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{4A7226CF-57BF-4CA3-A4AC-91A398A1D84B}"
 EndProject
@@ -41,7 +44,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TokenCredentialDF", "sample
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TokenCredentialDTFx", "samples\TokenCredentialDTFx\TokenCredentialDTFx.csproj", "{FBFF0814-E6C0-489A-ACCF-9D0699219621}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Functions.Worker.Extensions.DurableTask.Netherite", "src\Functions.Worker.Extensions.DurableTask.Netherite\Functions.Worker.Extensions.DurableTask.Netherite.csproj", "{3E17402B-3F65-4E5B-B752-48AD56B81208}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Functions.Worker.Extensions.DurableTask.Netherite", "src\Functions.Worker.Extensions.DurableTask.Netherite\Functions.Worker.Extensions.DurableTask.Netherite.csproj", "{3E17402B-3F65-4E5B-B752-48AD56B81208}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DurableTask.Netherite.EventHubs", "src\DurableTask.Netherite.EventHubs\DurableTask.Netherite.EventHubs.csproj", "{A8370E8C-11D2-4C24-BFB1-082F2D8F9890}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -97,6 +102,10 @@ Global
 		{3E17402B-3F65-4E5B-B752-48AD56B81208}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3E17402B-3F65-4E5B-B752-48AD56B81208}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3E17402B-3F65-4E5B-B752-48AD56B81208}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8370E8C-11D2-4C24-BFB1-082F2D8F9890}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8370E8C-11D2-4C24-BFB1-082F2D8F9890}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8370E8C-11D2-4C24-BFB1-082F2D8F9890}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8370E8C-11D2-4C24-BFB1-082F2D8F9890}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -114,6 +123,7 @@ Global
 		{B99AB043-47DC-467C-93CB-D6C69D3B1AD4} = {AB958467-9236-402E-833C-B8DE4841AB9F}
 		{FBFF0814-E6C0-489A-ACCF-9D0699219621} = {AB958467-9236-402E-833C-B8DE4841AB9F}
 		{3E17402B-3F65-4E5B-B752-48AD56B81208} = {D33AB157-04B9-4BAD-B580-C3C87C17828C}
+		{A8370E8C-11D2-4C24-BFB1-082F2D8F9890} = {D33AB157-04B9-4BAD-B580-C3C87C17828C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {238A9613-5411-41CF-BDEC-168CCD5C03FB}
