+++ src/DurableTask.Netherite.AzureFunctions/DurableTask.Netherite.AzureFunctions.csproj
@@ -1,62 +1,22 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
-	<!--The target netcoreapp2.2 is not functional, but just generates runtime error when used.-->
+	  <!--The target netcoreapp2.2 is not functional, but just generates runtime error when used.-->
     <TargetFrameworks>netstandard2.1;netstandard2.0;netcoreapp2.2;netcoreapp3.1</TargetFrameworks>
-    <LangVersion>latest</LangVersion>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <IncludeSymbols>true</IncludeSymbols>
-    <DebugType>embedded</DebugType>
-    <RepositoryUrl>https://github.com/microsoft/durabletask-netherite</RepositoryUrl>
-    <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <EmbedUntrackedSources>true</EmbedUntrackedSources>
-    <Authors>Microsoft</Authors>
-    <PackageProjectUrl>https://github.com/microsoft/durabletask-netherite</PackageProjectUrl>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Copyright>© Microsoft Corporation. All rights reserved.</Copyright>
-    <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>..\..\sign.snk</AssemblyOriginatorKeyFile>
+    
+    <PackageId>Microsoft.Azure.DurableTask.Netherite.AzureFunctions</PackageId>
     <Description>Netherite durability provider extension for Azure Durable Functions.</Description>
     <PackageTags>Azure Task Durable Orchestration Workflow Activity Reliable EventHubs</PackageTags>
-    <PackageId>Microsoft.Azure.DurableTask.Netherite.AzureFunctions</PackageId>
-    <PackageIcon>icon.png</PackageIcon>
+
+    <!-- Our netcoreapp2.2 target is a non-functional dummy target, so we don't need the warning -->    
+	  <NoWarn Condition="'$(TargetFramework)' == 'netcoreapp2.2'">NETSDK1138</NoWarn>
   </PropertyGroup>
 
-  <!-- Version settings: https://andrewlock.net/version-vs-versionsuffix-vs-packageversion-what-do-they-all-mean/ -->
-  <PropertyGroup>
-	<MajorVersion>1</MajorVersion>
-	<MinorVersion>4</MinorVersion>
-	<PatchVersion>0</PatchVersion>
-	<VersionPrefix>$(MajorVersion).$(MinorVersion).$(PatchVersion)</VersionPrefix>
-    <VersionSuffix></VersionSuffix>
-    <AssemblyVersion>$(MajorVersion).0.0.0</AssemblyVersion>
-    <BuildSuffix Condition="'$(GITHUB_RUN_NUMBER)' != ''">.$(GITHUB_RUN_NUMBER)</BuildSuffix>
-    <FileVersion>$(VersionPrefix)$(BuildSuffix)</FileVersion>
-  </PropertyGroup>
-
-  <!-- Our netcoreapp2.2 target is a non-functional dummy target, so we don't need the warning -->
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'netcoreapp2.2' ">
-	<NoWarn>NETSDK1138</NoWarn>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <None Include="icon.png" Pack="true" PackagePath="\" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(Configuration)'=='Release'">
-    <Content Include="..\..\_manifest\**">
-      <Pack>true</Pack>
-      <PackagePath>content/SBOM</PackagePath>
-    </Content>
-  </ItemGroup>
-  
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.DurableTask.Core" Version="2.13.0" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.DurableTask" Version="2.9.6" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFramework)' != 'netcoreapp2.2' ">
-	<ProjectReference Include="..\DurableTask.Netherite\DurableTask.Netherite.csproj" />
-  </ItemGroup>
-	
+  	<ProjectReference Include="..\DurableTask.Netherite\DurableTask.Netherite.csproj" />
+  </ItemGroup>	
 </Project>
