+++ src/DurableTask.Netherite.EventHubs/EventHubsUtil.cs
