+++ src/DurableTask.Netherite.EventHubs/TransportLayer/EventHubs/EventHubsConnections.cs
