+++ src/DurableTask.Netherite.EventHubs/TransportLayer/EventHubs/ScriptedEventProcessorHost.cs
