+++ src/DurableTask.Netherite/AssemblyInfo.cs
@@ -12,3 +12,9 @@
     + "0c6bc0e84740f74ca828fed9aebf15272867019cf991484c06b00c7d7aacb34b40ed0ae633f043"
     + "2df41db65caa4b03f9a0e6974c15aeaa7db2acd902914c0628f0b77d9b7839f530ba04584aed92"
     + "87159ebd")]
+[assembly: InternalsVisibleTo("DurableTask.Netherite.EventHubs, PublicKey="
+    + "0024000004800000940000000602000000240000525341310004000001000100e93024fd5f6737"
+    + "b5f408b39a97e746f673e7c62d690716c6022ee8871b2bc5efcbd8015dd02253302196ded0d9fc"
+    + "0c6bc0e84740f74ca828fed9aebf15272867019cf991484c06b00c7d7aacb34b40ed0ae633f043"
+    + "2df41db65caa4b03f9a0e6974c15aeaa7db2acd902914c0628f0b77d9b7839f530ba04584aed92"
+    + "87159ebd")]
