+++ src/DurableTask.Netherite/Scaling/ScalingMonitor.cs
@@ -11,7 +11,6 @@ namespace DurableTask.Netherite.Scaling
     using System.Text;
     using System.Threading;
     using System.Threading.Tasks;
-    using DurableTask.Netherite.EventHubsTransport;
 
     /// <summary>
     /// Monitors the performance of the Netherite backend and makes scaling decisions.
@@ -215,9 +214,7 @@ namespace DurableTask.Netherite.Scaling
 
             // next, check if any of the entries are not current, in the sense that their input queue position
             // does not match the latest queue position
-
-           
-            List<long> positions = await Netherite.EventHubsTransport.EventHubsConnections.GetQueuePositionsAsync(this.eventHubsConnection, EventHubsTransport.PartitionHub).ConfigureAwait(false);
+            List<long> positions = await EventHubsConnections.GetQueuePositionsAsync(this.eventHubsConnection, EventHubsTransport.PartitionHub).ConfigureAwait(false);
 
             if (positions == null)
             {
